// src/app/model/spp.model.ts
export interface Spp {
  sppId?: string;
  name?: string;
  address?: string;
  tel?: string;
  fax?: string;
  director?: string;
  status?: string;
  spcId?: string;
  spcName?: string;
  polId?: string;
  polName?: string;
  locaId?: string;
  position?: string;
  shortName?: string;
  locaName?: string;
  sppCode?: string;
  sppLevel?: string;
  sppParent?: string;
  oderCode?: string;
  isDePart?: string;
  sppIdFOX?: string;
  nameSync?: string;
}